Point the dfe-wizard gem to 0.1.0
Definitions of Rails routes prefix namespace for A levels with default path arguments

Of one example is the first step to the second step:

 publish_provider_recruitment_cycle_course_a_levels_what_a_level_is_required

 publish_provider_recruitment_cycle_course - defined below
 a_levels - ALevelSteps module
 what_a_level_is_required - WhatALevelIsRequired step

 this will produce the following url
 /publish/organisations/:provider_code/:recruitment_cycle_year/courses/:course_code
